+++ flaredb/src/engine/mod.rs
@@ -2,5 +2,6 @@ pub mod coders;
 pub mod dispatcher;
 pub mod executor;
 pub mod harness;
+pub mod runtime;
 pub mod scheduler;
 pub mod sdf;
